add authentication middleware

+++ src/app/article/[slug]/_components/commentCard/action.ts
@@ -1,10 +1,16 @@
 "use server";
 
 import { createApiClient } from "@/utils/api/apiClient";
+import { getSession } from "@/utils/auth/session";
 import { revalidateTag } from "next/cache";
+import { redirect } from "next/navigation";
 import { Inputs } from "./types";
 
 export const deleteCommentAction = async (_prevState: undefined, inputs: Inputs): Promise<undefined> => {
+  if ((await getSession()) == null) {
+    redirect("/login");
+  }
+
   const apiClient = createApiClient({
     path: "/articles/{slug}/comments/{id}",
     method: "delete",
